Add commentary about leaflet
Added commentary about some of the limitations experienced using the leaflet package to visualize the data. These limitations were due to overall unfamiliarity with the package, and shortness of time to fully solve the experienced problems. The package can be used in the future to better visualize the data, and allow DCLS values to be viewed for the different trip purposes using choropleth maps at the same time. Additionally, these leaflet choropleths are much  nicer than the others generated in the analysis.